feat: JS modules: use esm to provide both CommonJS and ES6 entry points
This also adds standard-version to our development process.

This module's function is now under .default for CommonJS users.
