+++ README.md
@@ -8,6 +8,13 @@
 Calculate the [relative luminance](http://www.w3.org/TR/2008/REC-WCAG20-20081211/#relativeluminancedef) of
 an RGB triplet color.
 
+Note that as of 1.0.0, this module doesn't directly export a function if you're using
+CommonJS: use the `default` export, like:
+
+```js
+var luminance = require('relative-luminance').default;
+```
+
 ### API
 
 <!-- Generated by documentation.js. Update this documentation by updating the source code. -->
@@ -24,7 +31,7 @@ as a number from 0 to 1.
 **Examples**
 
 ```javascript
-var luminance = require('relative-luminance');
+var luminance = require('relative-luminance').default;
 var black_lum = luminance([0, 0, 0]); // 0
 ```
